+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add DevMode panel (trigger: CMD+') - #3333 by @zaiste 
 - Migrate to `es2020` for TypeScript - #3386 by @zaiste 
 - Fix styling for GraphiQL on the webhook page - #3389 by @zaiste
+- Copy filters automatically to GraphiQL playground - #3385 by @zaiste
 
 ## 3.4